refactor(hooks): migrate remaining hooks to TypeScript
Migrate the last 3 JavaScript hooks to TypeScript:
- useLongPress.js → useLongPress.ts
- usePullToRefresh.js → usePullToRefresh.ts
- useSwipeNavigation.js → useSwipeNavigation.ts

All hooks now have:
- Full TypeScript type definitions
- Exported interfaces for options and return types
- TSDoc documentation
- Proper React event typing

Update index.ts to export new types:
- LongPressOptions, LongPressHandlers
- PullToRefreshOptions, PullToRefreshHandlers, PullToRefreshReturn
- SwipeNavigationOptions, SwipeNavigationHandlers, SwipeNavigationReturn, SwipeDirection

The src/hooks/ directory is now 100% TypeScript.
